refactor: collocation
